ref(onboarding): Split aws lambda onboarding docs (#102945)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
