+++ static/app/gettingStartedDocs/dotnet/awslambda/index.tsx
@@ -1,0 +1,13 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {feedback} from 'sentry/gettingStartedDocs/dotnet/dotnet/feedback';
+
+import {crashReport} from './crashReport';
+import {onboarding} from './onboarding';
+
+const docs: Docs = {
+  onboarding,
+  feedbackOnboardingCrashApi: feedback,
+  crashReportOnboarding: crashReport,
+};
+
+export default docs;
