+++ static/app/gettingStartedDocs/dotnet/awslambda/onboarding.spec.tsx
@@ -4,7 +4,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './awslambda';
+import docs from './index';
 
 describe('awslambda onboarding docs', () => {
   it('renders errors onboarding docs correctly', async () => {
